train_predict_service/app/requirements.txt
et-xmlfile==1.1.0
mysql-connector-python==8.0.26
nest-asyncio==1.5.1
numpy==1.19.5
openpyxl==3.0.9
pandas==1.1.5
pickleshare==0.7.5
xlrd==2.0.1
sklearn
nltk
flask
bs4==0.0.1
requests==2.26.0
kafka_python==2.0.2
pyspark
sqlalchemy==1.4.25
mysqlclient

